+++ index.d.ts
@@ -14,6 +14,7 @@
 import { ChainLib, ChainConfig } from './typings/chain';
 import { MaximusChainLib } from './typings/chains/maximus';
 import { OsmiumChainLib } from './typings/chains/osmium';
+import { FilopowChainLib } from './typings/chains/filopow';
 
 /** Library version, e.g. `"v3.0.2"`. */
 export const version: string;
@@ -47,6 +48,7 @@ export function createHashRegistry(): import('./typings/chain').HashRegistry;
  */
 export function create(name: 'maximus'): MaximusChainLib;
 export function create(name: 'osmium'): OsmiumChainLib;
+export function create(name: 'filopow'): FilopowChainLib;
 export function create(name: string): ChainLib;
 
 /**